implement creating the schema element table
